Dynamic UVC controls are taken from the guvcview 1.1.3 which is said to be working.
git-svn-id: https://mjpg-streamer.svn.sourceforge.net/svnroot/mjpg-streamer@115 a8dad676-403b-0410-b964-c43bdf40fe89

+++ mjpg-streamer/plugins/output_http/httpd.c
@@ -401,7 +401,7 @@ void send_stream(int fd) {
     sprintf(buffer, "Content-Type: image/jpeg\r\n" \
                     "Content-Length: %d\r\n" \
                     "X-Timestamp: %d.%06d\r\n" \
-                    "\r\n", frame_size, timestamp.tv_sec, timestamp.tv_usec);
+                    "\r\n", frame_size, (int)timestamp.tv_sec, (int)timestamp.tv_usec);
     DBG("sending intemdiate header\n");
     if ( write(fd, buffer, strlen(buffer)) < 0 ) break;
